Refactor view globals to allow use of request
