+++ modules/viewGlobals.test.js
@@ -1,28 +1,13 @@
 /* global describe, it */
+'use strict';
 
 const chai = require('chai');
 const expect = chai.expect;
 
 const httpMocks = require('node-mocks-http');
-const { buildUrl, getCurrentUrl, getMetaTitle, getCurrentSection } = require('./viewGlobals');
+const { getCurrentUrl, getMetaTitle, getCurrentSection } = require('./viewGlobals');
 
 describe('View Globals', () => {
-    describe('#buildUrl', () => {
-        it('should build correct url based on section url and page name', () => {
-            const builderEn = buildUrl('');
-            const builderCy = buildUrl('welsh');
-
-            expect(builderEn('toplevel', 'benefits')).to.equal('/jobs/benefits');
-            expect(builderCy('toplevel', 'benefits')).to.equal('welsh/jobs/benefits');
-        });
-
-        it('should build correct url when a simple path is given', () => {
-            const testPath = 'global-content/programmes/england/awards-for-all-england';
-            expect(buildUrl('')(testPath)).to.equal(`/${testPath}`);
-            expect(buildUrl('welsh')(testPath)).to.equal(`welsh/${testPath}`);
-        });
-    });
-
     describe('#getCurrentUrl', () => {
         it('should return expected url for en locale', () => {
             const req = httpMocks.createRequest({
